app/requirements.txt
telegram==0.0.1
async-timeout==3.0.1
proxybroker==0.3.2
python-telegram-bot==12.4.2
requests==2.23.0
python-dotenv==0.12.0